Stylized macOS + sorted Desktop & Mobile OS names
Since iOS was stylized, I thought it was consistent if we also stylized macOS (it's not Mac OS anymore, it's (macOS)[https://en.wikipedia.org/wiki/MacOS]). 

Also took a quick moment to sort the values for Desktop and Mobile OS names in order to more easily locate at a glance the item being looked for.